requirements.txt
matplotlib==3.6.2
numpy==1.19.5
pandas==1.2.4
Pillow==9.3.0
plotly==5.9.0
requests==2.25.1
streamlit==1.14.1
